modification d'accesbddcolis avec des float pour le volume

+++ trunk/accesBDD/AccesBDDColis.java
@@ -43,7 +43,7 @@ public class AccesBDDColis extends AccesBDD{
 		ajout.setTimestamp(9, aAjouter.getDate());
 		ajout.setString(10, aAjouter.getValeurDeclaree());
 		ajout.setInt(11, aAjouter.getFragilite().intValue());
-		ajout.setInt(12, aAjouter.getVolume().intValue());
+		ajout.setFloat(12, aAjouter.getVolume().floatValue());
 		ajout.setInt(13, aAjouter.getOrigine().getId().intValue());
 		ajout.setInt(14, aAjouter.getEntrepot().getId().intValue());
 		
@@ -89,7 +89,7 @@ public class AccesBDDColis extends AccesBDD{
 		modifie.setTimestamp(8, aModifier.getDate());
 		modifie.setString(9, aModifier.getValeurDeclaree());
 		modifie.setInt(10, aModifier.getFragilite().intValue());
-		modifie.setInt(11, aModifier.getVolume().intValue());
+		modifie.setFloat(11, aModifier.getVolume().floatValue());
 		modifie.setInt(12,aModifier.getId().intValue());
 		modifie.setInt(13, aModifier.getOrigine().getId().intValue());
 		modifie.setInt(14, aModifier.getEntrepot().getId().intValue());
